# apigen/model/openapi.yaml
openapi: 3.0.3
info:
  title: Data Repository Service
  version: 1.5.0
  x-logo:
    url: https://www.ga4gh.org/wp-content/themes/ga4gh/dist/assets/svg/logos/logo-full-color.svg
  termsOfService: https://www.ga4gh.org/terms-and-conditions/
  contact:
    name: GA4GH Cloud Work Stream
    email: ga4gh-cloud@ga4gh.org
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/ga4gh/data-repository-service-schemas/master/LICENSE
servers:
  - url: https://{serverURL}/ga4gh/drs/v1
    variables:
      serverURL:
        default: drs.example.org
        description: |
          DRS server endpoints MUST be prefixed by the '/ga4gh/drs/v1' endpoint path
security:
  - {}
  - BasicAuth: []
  - BearerAuth: []
...
paths:
  /service-info:
    get:
      summary: Retrieve information about this service
      description: "Returns information about the DRS service along with stats pertaning to total object count and cumulative size in bytes.\nAlso indicates whether the server supports optional upload and delete operations and which methods are available.\n\nExtends the\n[v1.0.0 GA4GH Service Info specification](https://github.com/ga4gh-discovery/ga4gh-service-info)\nas the standardized format for GA4GH web services to self-describe.\n\nAccording to the \n[service-info type registry](https://github.com/ga4gh/TASC/blob/master/service-info/ga4gh-service-info.json)\nmaintained by the [Technical Alignment Sub Committee (TASC)](https://github.com/ga4gh/TASC),\na DRS service MUST have:\n  * a `type.group` value of `org.ga4gh`\n  * a `type.artifact` value of `drs`\n\n**Example 1: Server with upload and delete capabilities**\n```\n{\n    \"id\": \"com.example.drs\",\n    \"description\": \"Serves data according to DRS specification\",\n    ...\n    \"type\": {\n        \"group\": \"org.ga4gh\",\n        \"artifact\": \"drs\",\n        \"version\": \"1.5\"\n    }\n    ...\n    \"drs\":{\n      \"maxBulkRequestLength\": 200,\n      \"objectCount\": 774560,\n      \"totalObjectSize\": 4018437188907752,\n      \"uploadRequestSupported\": true,\n      \"objectRegistrationSupported\": true,\n      \"supportedUploadMethodTypes\": [\"s3\", \"https\", \"gs\"],\n      \"maxUploadSize\": 5368709120,\n      \"maxUploadRequestLength\": 50,\n      \"validateChecksums\": true,\n      \"validateFileSizes\": false,\n      \"relatedFileStorageSupported\": true,\n      \"deleteSupported\": true,\n      \"maxBulkDeleteLength\": 100,\n      \"deleteStorageDataSupported\": true\n    }\n}\n```\n\n**Example 2: Read-only server (no upload or delete)**\n```\n{\n    \"id\": \"com.example.readonly-drs\",\n    \"description\": \"Read-only DRS service\",\n    ...\n    \"type\": {\n        \"group\": \"org.ga4gh\",\n        \"artifact\": \"drs\",\n        \"version\": \"1.5\"\n    }\n    ...\n    \"drs\":{\n      \"maxBulkRequestLength\": 500,\n      \"objectCount\": 1250000,\n      \"totalObjectSize\": 8500000000000000\n    }\n}\n```\n\n**Example 3: Server with metadata-only delete capability**\n```\n{\n    \"id\": \"com.example.metadata-drs\",\n    \"description\": \"DRS service with metadata-only delete\",\n    ...\n    \"type\": {\n        \"group\": \"org.ga4gh\",\n        \"artifact\": \"drs\",\n        \"version\": \"1.5\"\n    }\n    ...\n    \"drs\":{\n      \"maxBulkRequestLength\": 200,\n      \"objectCount\": 500000,\n      \"totalObjectSize\": 2500000000000000,\n      \"deleteSupported\": true,\n      \"maxBulkDeleteLength\": 50,\n      \"deleteStorageDataSupported\": false\n    }\n}\n```\nA service that supports object-level controlled-access claims SHOULD set `drs.controlledAccessSupported` to `true`, SHOULD set `drs.controlledAccessClaimFormat` to `ga4gh-passport-url-claim`, and SHOULD set `drs.controlledAccessDefault` to `open-access-read`.\n\nSee the [Service Registry Appendix](#tag/GA4GH-Service-Registry) for more information on how to register a DRS service with a service registry."
      operationId: GetServiceInfo
      responses:
        '200':
          $ref: '#/components/responses/200ServiceInfo'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Service Info
  /objects/{object_id}:
    options:
      summary: Get Authorization info about a DrsObject.
      security:
        - {}
      description: Returns a list of `Authorizations` that can be used to determine how to authorize requests to `GetObject` or `PostObject`.
      operationId: OptionsObject
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          $ref: '#/components/responses/200OkAuthorizations'
        '204':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '405':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
    get:
      summary: Get info about a DrsObject.
      description: Returns object metadata, and a list of access methods that can be used to fetch object bytes.
      operationId: GetObject
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/Expand'
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObject'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
    post:
      summary: Get info about a DrsObject through POST'ing a Passport.
      description: |-
        Returns object metadata and a list of access methods that can be used to fetch object bytes. Method is a POST to accommodate a JWT GA4GH Passport sent in the request body in order to authorize access.
        **Note**: To upload new files and register them as DRS objects, use the `/upload-request` endpoint  to obtain upload methods and temporary credentials, then use POST `/objects/register` endpoint  to register multiple objects at once. Note that upload functionality is optional and not all DRS  servers implement the upload endpoints.
      operationId: PostObject
      security:
        - PassportAuth: []
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObject'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundAccess'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      requestBody:
        $ref: '#/components/requestBodies/PostObjectBody'
  /objects:
    options:
      summary: Get Authorization info about multiple DrsObjects.
      security:
        - {}
      description: Returns a structure that contains for each DrsObjects a list of `Authorizations` that can be used to determine how to authorize requests to `GetObject` or `PostObject` (or bulk equivalents).
      operationId: OptionsBulkObject
      responses:
        '200':
          $ref: '#/components/responses/200OkBulkAuthorizations'
        '204':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '405':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '413':
          $ref: '#/components/responses/413RequestTooLarge'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
    post:
      summary: Get info about multiple DrsObjects with an optional Passport(s).
      description: |-
        Returns an array of object metadata and access methods for the specified object IDs.
        The request is limited to use passports (one or more) or a single bearer token,  so make sure your bulk request is for objects that all use the same passports/token.
        **Note**: To register new DRS objects, use the dedicated `/objects/register` endpoint.
      operationId: GetBulkObjects
      security:
        - PassportAuth: []
      parameters:
        - $ref: '#/components/parameters/Expand'
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObjects'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '413':
          $ref: '#/components/responses/413RequestTooLarge'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      requestBody:
        $ref: '#/components/requestBodies/BulkObjectBody'
  /objects/{object_id}/access/{access_id}:
    get:
      summary: Get a URL for fetching bytes
      description: |-
        Returns a URL that can be used to fetch the bytes of a `DrsObject`.
        This method only needs to be called when using an `AccessMethod` that contains an `access_id` (e.g., for servers that use signed URLs for fetching object bytes).
      operationId: GetAccessURL
      responses:
        '200':
          $ref: '#/components/responses/200OkAccess'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundAccess'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/AccessId'
    post:
      summary: Get a URL for fetching bytes through POST'ing a Passport
      description: |-
        Returns a URL that can be used to fetch the bytes of a `DrsObject`.
        This method only needs to be called when using an `AccessMethod` that contains an `access_id` (e.g., for servers that use signed URLs for fetching object bytes).
        Method is a POST to accommodate a JWT GA4GH Passport sent in the formData in order to authorize access.
      operationId: PostAccessURL
      security:
        - PassportAuth: []
      responses:
        '200':
          $ref: '#/components/responses/200OkAccess'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundAccess'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/AccessId'
      requestBody:
        $ref: '#/components/requestBodies/Passports'
  /objects/access:
    post:
      summary: Get URLs for fetching bytes from multiple objects with an optional Passport(s).
      description: |-
        Returns an array of URL objects that can be used to fetch the bytes of multiple `DrsObject`s.
        This method only needs to be called when using an `AccessMethod` that contains an `access_id` (e.g., for servers that use signed URLs for fetching object bytes).
        Currently this is limited to use passports (one or more) or a single bearer token, so make sure your bulk request is for objects that all use the same passports/token.
      operationId: GetBulkAccessURL
      security:
        - PassportAuth: []
      responses:
        '200':
          $ref: '#/components/responses/200OkAccesses'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundAccess'
        '413':
          $ref: '#/components/responses/413RequestTooLarge'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkObjectAccessId'
  /objects/register:
    post:
      summary: Register DRS objects
      description: "**Optional Endpoint**: This endpoint is not required for DRS server implementations.  Not all DRS servers support object registration.    \nRegisters one or more \"candidate\" DRS objects with the server. If it accepts the request, the server will create  unique object IDs for each registered object and return them in fully-formed DRS objects in response.\nThis endpoint can be used after uploading files using methods negotiated with the `/upload-request` endpoint  to register the uploaded files as DRS objects, or to register existing data. The request body should contain candidate  DRS objects with all required metadata including access methods that correspond  to the upload methods used during file upload.\n**RECOMMENDED - Transactional Behavior**:  Registration operations SHOULD be atomic transactions. If ANY candidate object fails validation  or registration, the ENTIRE request SHOULD fail and NO objects SHOULD be registered. Servers SHOULD  implement this as an all-or-nothing operation to ensure data consistency, but MAY implement partial  registration with appropriate error reporting if transactional behavior is not feasible.\n**Authentication**: GA4GH Passports can be provided in the request body for authorization. Bearer tokens can be supplied in headers.\n**Server Responsibilities**: - SHOULD treat registration as an atomic transaction (all succeed or all fail) - SHOULD validate ALL candidate objects before registering ANY - Create unique object IDs for each registered object - Add timestamps (created_time, updated_time) - SHOULD roll back any partial changes if any candidate fails validation\n**Client Responsibilities**: - Provide required DRS object metadata for all candidates - Include access methods corresponding to uploaded file locations - Ensure checksums match uploaded file content - Handle potential failure of entire batch if any single object is invalid"
      operationId: RegisterObjects
      security:
        - {}
        - BasicAuth: []
        - BearerAuth: []
        - PassportAuth: []
      requestBody:
        $ref: '#/components/requestBodies/RegisterObjectsBody'
      responses:
        '201':
          $ref: '#/components/responses/201ObjectsCreated'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '413':
          $ref: '#/components/responses/413RequestTooLarge'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Register Objects
      x-swagger-router-controller: ga4gh.drs.server
      x-codegen-request-body-name: body
...
  /objects/checksum/{checksum}:
    get:
      summary: Get DRS objects that are a match for the checksum.
      description: |-
        Returns an array of `DRSObjects` that match a given checksum.
        The checksum type is not provide, the checksum check is done against all checksum types.
      operationId: GetObjectsByChecksum
      security:
        - PassportAuth: []
      parameters:
        - $ref: '#/components/parameters/Checksum'
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObjects'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: |
        A valid authorization token must be passed in the 'Authorization' header,
        e.g. "Basic ${token_string}"
    BearerAuth:
      type: http
      scheme: bearer
      description: A valid authorization token must be passed in the 'Authorization' header, e.g. "Bearer ${token_string}"
    PassportAuth:
      type: http
      scheme: bearer
      x-in: body
      bearerFormat: JWT
      description: A valid GA4GH Passport must be passed in the body of an HTTP POST request as a tokens[] array.
...
  responses:
    200ServiceInfo:
      description: Retrieve info about the DRS service
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Service'
              - $ref: '#/components/schemas/DrsService'
    500InternalServerError:
      description: An unexpected error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    400BadRequest:
      description: The request is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    200OkAuthorizations:
      description: '`Authorizations` were found successfully'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Authorizations'
    AuthorizationsNotSupported:
      description: '`Authorizations` are not supported for this object. Default to `None`.'
...
  parameters:
    ObjectId:
      in: path
      name: object_id
      required: true
      description: '`DrsObject` identifier'
      schema:
        type: string
...